add & fix: 在 xoj/single_judge.go 中添加 spj 支持, 修复读取样例以及用户输出后有乱码的问题

+++ xoj/dao/judge_status.go
@@ -15,7 +15,7 @@ type JudgeStatus struct {
 	SubmitTime     time.Time
 	Judger         string
 	Contest        int64
-	ContestProblem int64
+	ContestProblem int
 	Code           string
 	TestCase       string
 	Message        string
